fix : 전문가, 관련 정보 엔티티 수정
