+++ src/main/java/dbdip/demo/expert/entity/Stores.java
@@ -1,4 +1,4 @@
-package dbdip.demo.expert;
+package dbdip.demo.expert.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -11,18 +11,20 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 @NoArgsConstructor(force = true)
 @Getter
-@Table(name = "Consult")
-public class Consult {
+public class Stores {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "cid")
-    private int id;
-    @Column(name = "field")
-    private String field;
-    @Column(name = "price_hour")
-    private Double priceHour;
+    @Column(name = "sid")
+    private Integer id;
+    @Column(name = "phonenumber")
+    private String phoneNumber;
+    @Column(name = "city")
+    private String city;
+    @Column(name = "zipcode")
+    private String zipCode;
 
     @OneToOne
     @JoinColumn(name = "eid")
     private Experts experts;
+
 }
